Set default bedding to single
Update AddBookingModal to initialize bedding as single by default in hotel allocations, ensuring single bedding is selected unless user changes it. This aligns with readme requiring deliberate choice of bedding type.

+++ src/components/AddBookingModal.tsx
@@ -134,7 +134,7 @@ export const AddBookingModal = ({ open, onOpenChange, preSelectedTourId, default
           allocated: formData.accommodation_required,
           check_in_date: hotel.default_check_in || preSelectedTourStartDate || '',
           check_out_date: hotel.default_check_out || preSelectedTourEndDate || '',
-          bedding: 'double',
+          bedding: 'single',
           room_type: hotel.default_room_type || '',
           room_upgrade: '',
           confirmation_number: '',
